requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-PyMongo==2.3.0
Werkzeug==2.3.7
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
openpyxl==3.1.2
reportlab==4.0.4
Pillow==10.0.0
bcrypt==4.0.1
celery==5.3.1
redis==5.0.0
gunicorn==21.2.0
googletrans==4.0.0-rc1